Equipment filters on geoJSON API.
