+++ static/scss/style.scss
@@ -1146,7 +1146,6 @@ select[name='language'] {
   background-position: 50% !important;
 }
 
-
 div.equipments {
   margin: 15px;
 }
